Removed all Fog actors in map
Deleted all 189 Things containing a ThingID of '15001' (Fog).  This was removed as it potentially caused negated performance when uzing Zandronum's 3.0 OpenGL renderer engine.  All assets related to the 'Fog' feature will remain as-is if others decide to use it.

I was not able to test it as Zandronum throws an abundance of DECORATE errors when I try to run the whole project.  Someone more experienced with this project will need to perform a test to assure that there is (or is not) an performance improvement.  If not, the next phase would be to either: Reduce the rain actors by 50% and virtual grow the rain actors (which the players may never see the rain particles), or expunge the entire rain and the rain ambience entirely.

>> NEEDS TESTING