WIP - Monsters works from LOADACS
The monster waves appear to tbe working by using the invasion mode as a library. I still have to do the items on that part, though. Shouldn't be a big deal to do but I'm out of time right now.
There are currently a lot of debugging messages running, which of course won't be there in the final version.
Monsters spawning other monsters works too. There is a pain elemental for a reason. ;)

I don't know how mapping with libraries works, I just assume it's doable as TEHACS.acs seems to be used from the maps as an import. Including IMONLIST.acs is a good idea as well because it contains definitions for indexes into the monster name string table, just so things are a bit more readable when adding those 100 lines of wave group registrations.

For an example on the usage so far, just look into the TESTMAP map, as it contains an OPEN script demonstrating the setup.

Further development:
> Item registration from map OPEN script.
> Have the map handle the waves. (Should be as simple as moving that script from the INVMODE.acs file to the map but I have to do it to confirm that)
> In general a lot of things I have forgotten the specifics of right now.